move operator display functions
